+++ addons/pvr.vdr.vnsi/vdr-plugin-vnsiserver/videobuffer.h
@@ -35,11 +35,11 @@ public:
   static cVideoBuffer* Create(cRecording *rec);
   virtual void Put(uint8_t *buf, unsigned int size) = 0;
   virtual int ReadBlock(uint8_t **buf, unsigned int size) = 0;
-  virtual size_t GetPosMin() { return 0; };
-  virtual size_t GetPosMax() { return 0; };
-  virtual size_t GetPosCur() { return 0; };
-  virtual void GetPositions(size_t *cur, size_t *min,size_t *max) {};
-  virtual void SetPos(size_t pos) {};
+  virtual off_t GetPosMin() { return 0; };
+  virtual off_t GetPosMax() { return 0; };
+  virtual off_t GetPosCur() { return 0; };
+  virtual void GetPositions(off_t *cur, off_t *min, off_t *max) {};
+  virtual void SetPos(off_t pos) {};
   virtual void SetCache(bool on) {};
   virtual bool HasBuffer() { return false; };
   int Read(uint8_t **buf, unsigned int size);
